+++ tea-coffee-core/event-page.php
@@ -6,7 +6,7 @@
  *
  * @package Wordpress
  * @subpackage Tea Coffee
- * @version 1.2.3
+ * @version 1.2.4
  * 
  * Template Name: Event
  */