+++ tea-coffee-core/front-page.php
@@ -6,7 +6,7 @@
  *
  * @package Wordpress
  * @subpackage Tea Coffee
- * @version 1.2.3
+ * @version 1.2.4
  */
 
 get_template_part( 'template-parts/site/site', 'head' ); ?>