+++ tea-coffee-core/functions.php
@@ -6,7 +6,7 @@
  *
  * @package Wordpress
  * @subpackage Tea Coffee
- * @version 1.2.3
+ * @version 1.2.4
  */
 
 /**
@@ -124,7 +124,7 @@ function cfe_splash_info( $wp_customize ) {
     )));
 
     $wp_customize->add_setting('cfe-splash-general-info', array(
-        'default' => 'Drop by Rm 5370, SFU Surrey Campus every Monday for our work meetings'
+        'default' => 'Info'
     ));
     $wp_customize->add_control( new WP_Customize_Control($wp_customize, '
     cfe-splash-general-info-control', array(