+++ tea-coffee-core/template-parts/post/content-featured.php
@@ -6,7 +6,7 @@
  *
  * @package Wordpress
  * @subpackage Tea Coffee
- * @version 1.2.3
+ * @version 1.2.4
  */
 
 $sticky = get_option( 'sticky_posts' );