+++ CHANGELOG.md
@@ -85,3 +85,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 [1.2.2]: https://github.com/technovus-sfu/technovus-theme/compare/v1.2.1...v1.2.2
 [1.2.1]: https://github.com/technovus-sfu/technovus-theme/compare/v1.2.0...v1.2.1
 [1.2.0]: https://github.com/technovus-sfu/technovus-theme/compare/v1.0.0...v1.2.0
+
+
+[101]: https://github.com/technovus-sfu/tea-coffee/commit/
+[100]: https://github.com/technovus-sfu/tea-coffee/commit/